fix(cmake): build script reviewed to improved compilation on centos8

+++ cmake.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "$1" = "-f" ] ; then
+  force=1
+  shift
+fi
+
 v=$(cmake --version)
 if [[ $v =~ "version 3" ]] ; then
   cmake='cmake'
@@ -33,9 +38,19 @@ else
   echo "'build' directory already there"
 fi
 
-rm -rf build
-mkdir build
+if [ $force -eq 1 ] ; then
+  rm -rf build
+  mkdir build
+fi
 cd build
-conan install .. --remote centreon
+
+maj=$(cat /etc/centos-release | awk '{print $4}' | cut -f1 -d'.')
+if [ $maj -eq "8" ] ; then
+  conan install .. --remote centreon -s compiler.libcxx=libstdc++11
+else
+  conan install .. --remote centreon -s compiler.libcxx=libstdc++
+fi
+
+echo "$*"
 
 CXXFLAGS="-Wall -Wextra" $cmake -DCMAKE_BUILD_TYPE=Debug -DWITH_PREFIX=/usr -DWITH_PREFIX_BIN=/usr/sbin -DWITH_USER=centreon-broker -DWITH_GROUP=centreon-broker -DWITH_CONFIG_PREFIX=/etc/centreon-broker -DWITH_TESTING=On -DWITH_PREFIX_MODULES=/usr/share/centreon/lib/centreon-broker -DWITH_PREFIX_CONF=/etc/centreon-broker -DWITH_PREFIX_LIB=/usr/lib64/nagios -DWITH_MODULE_SIMU=On $* ..
